refactor: 공간 등록 페이지 리팩토링 및 폴더 정리
