+++ src/components/button/Button.tsx
@@ -1,17 +1,30 @@
 type ButtonProps = {
   children: React.ReactNode;
   isValid: boolean;
+  disabled?: boolean;
+  onClick?: () => void;
+  type?: "button" | "submit" | "reset";
+  className?: string;
 };
 
-const Button = ({ children, isValid }: ButtonProps) => {
+const Button = ({
+  children,
+  isValid,
+  disabled = false,
+  onClick,
+  type = "submit",
+  className = "",
+}: ButtonProps) => {
   return (
     <button
-      type="submit"
-      className={`h-[5rem] w-full rounded-[1rem] text-16-Medium transition-colors duration-200 ${
-        isValid
-          ? "bg-[#000] text-white"
-          : "bg-[#D9D9D9] text-black cursor-not-allowed"
-      }`}
+      type={type}
+      disabled={disabled || !isValid}
+      onClick={onClick}
+      className={`h-[5.6rem] w-full rounded-[1rem] text-16-SemiBold transition-colors duration-200 ${
+        isValid && !disabled
+          ? "bg-cr-black text-cr-white"
+          : "bg-cr-200 text-cr-600 cursor-not-allowed"
+      } ${className}`}
     >
       {children}
     </button>
