+++ src/pages/host-spaceregister/hostspaceregisterinfo-components/input/UseCategory.tsx
@@ -1,0 +1,104 @@
+import { useState, useEffect, useRef } from "react";
+import type {
+  UseFormRegister,
+  UseFormSetValue,
+  FieldErrors,
+} from "react-hook-form";
+import { warning, topArrow, underArrow } from "../../../../assets/theme";
+import type { HostSpaceInfo } from "../../../../types/HostSpaceInfo";
+
+interface UseCategoryProps {
+  register: UseFormRegister<HostSpaceInfo>;
+  setValue: UseFormSetValue<HostSpaceInfo>;
+  errors: FieldErrors<HostSpaceInfo>;
+}
+
+const UseCategory = ({ register, setValue, errors }: UseCategoryProps) => {
+  const [isDropdownOpen, setIsDropdownOpen] = useState(false);
+  const [selectedPurpose, setSelectedPurpose] = useState("");
+  const dropdownRef = useRef<HTMLDivElement>(null);
+
+  const purposeList = [
+    { value: "MEETING", label: "미팅" },
+    { value: "COOKING", label: "쿠킹" },
+    { value: "BARISTA", label: "바리스타" },
+    { value: "FLEA_MARKET", label: "플리마켓" },
+    { value: "FILMING", label: "촬영" },
+    { value: "ETC", label: "기타" },
+  ] as const;
+
+  // 드롭다운 외부 클릭 시 닫기
+  useEffect(() => {
+    const handleClickOutside = (event: MouseEvent) => {
+      if (
+        dropdownRef.current &&
+        !dropdownRef.current.contains(event.target as Node)
+      ) {
+        setIsDropdownOpen(false);
+      }
+    };
+
+    document.addEventListener("mousedown", handleClickOutside);
+    return () => document.removeEventListener("mousedown", handleClickOutside);
+  }, []);
+
+  return (
+    <div className="flex flex-col gap-[0.8rem]">
+      <label className="text-13-SemiBold">
+        사용 용도<span className="text-cr-red">*</span>
+      </label>
+
+      <div className="relative" ref={dropdownRef}>
+        <button
+          type="button"
+          className={`flex justify-between items-center w-full h-[5.6rem] bg-cr-100 rounded-[1rem] text-14-Medium text-left px-[1.7rem] ${
+            selectedPurpose ? "text-cr-black" : "text-cr-500"
+          } ${errors.useCategory ? "border-2 border-red-500" : ""}`}
+          onClick={() => setIsDropdownOpen((prev) => !prev)}
+        >
+          <span>
+            {selectedPurpose
+              ? purposeList.find((p) => p.value === selectedPurpose)?.label
+              : "카테고리를 선택해주세요"}
+          </span>
+          <img src={isDropdownOpen ? topArrow : underArrow} alt="화살표" />
+        </button>
+
+        {isDropdownOpen && (
+          <ul className="z-10 bg-cr-white absolute mt-[0.8rem] w-full border-cr-300 border rounded-[1rem] px-[1.6rem] py-[1.2rem]">
+            {purposeList.map((purpose) => (
+              <li
+                key={purpose.value}
+                onClick={() => {
+                  setSelectedPurpose(purpose.value);
+                  setValue("useCategory", purpose.value);
+                  setIsDropdownOpen(false);
+                }}
+                className="h-[3rem] text-14-Medium flex items-center cursor-pointer hover:bg-cr-100 rounded-[0.5rem] px-[0.5rem]"
+              >
+                {purpose.label}
+              </li>
+            ))}
+          </ul>
+        )}
+
+        <input
+          type="hidden"
+          value={selectedPurpose}
+          {...register("useCategory")}
+        />
+      </div>
+
+      {errors.useCategory && (
+        <div className="flex gap-[0.6rem] items-center">
+          <img src={warning} alt="경고 아이콘" />
+          <span className="text-12-Medium text-red-500">
+            {errors.useCategory.message}
+          </span>
+        </div>
+      )}
+    </div>
+  );
+};
+
+export default UseCategory;
